+++ src/e2e/java/uk.gov.hmcts.reform.sscs/ProcessFileAndSaveIntoCcd.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import uk.gov.hmcts.reform.sscs.services.service.CaseLoaderService;
+import uk.gov.hmcts.reform.sscs.services.CaseLoaderService;
 
 /**
  * Class for testing purpose.
